Stop running stress-ng tests as root
The only stressor that won't run is apparmor, but it's non fatal and skipped.
Running as root in the other hand fails the tests on some platforms,
cpu and scheduler classes:

stress-ng: info: [30363] stress-ng-exec: running as root, won't run test.
stress-ng: error: [7373] process 30363 (stress-ng-exec) terminated with an error, exit status=1